AppData model and Projectstructure improvements
