+++ src/lib/mod.rs
@@ -1,0 +1,2 @@
+pub mod gui;
+pub mod model;
